Move cargo_rustc::compile_targets() to Context::compile()
