+++ src/cargo/ops/mod.rs
@@ -2,7 +2,7 @@ pub use self::cargo_clean::{clean, CleanOptions};
 pub use self::cargo_compile::{compile, compile_with_exec, compile_ws, CompileOptions};
 pub use self::cargo_compile::{CompileFilter, CompileMode, FilterRule, MessageFormat, Packages};
 pub use self::cargo_read_manifest::{read_package, read_packages};
-pub use self::cargo_rustc::{compile_targets, Compilation, Kind, Unit};
+pub use self::cargo_rustc::{Compilation, Kind, Unit};
 pub use self::cargo_rustc::{is_bad_artifact_name, Context};
 pub use self::cargo_rustc::{BuildConfig, BuildOutput, TargetConfig};
 pub use self::cargo_rustc::{DefaultExecutor, Executor};
